refactor(dashboard): reestructuración y modularización de la vista Dashboard
